Minor sidebar HilIndex and redraw cleanup
Ensure HilIndex is reset to -1 if all mailboxes are hidden, preventing
<sidebar-open> from opening a hidden mailbox in that case.

Previously, draw_sidebar() would still be triggered in that case,
causing an empty loop to run and trailing space to be drawn.  But with
this change TopIndex and BottomIndex will still be -1 in that case.

Change the prepare_sidebar() return value to check if HilIndex is set,
and if mutt_sb_draw() to clear out in that case.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/1e9ab8b67e50f512bed61f204cdc99c6ca76ba54
Co-authored-by: Richard Russon <rich@flatcap.org>