Change disabled crypto pattern op error message and return value
If mutt is compiled without crypto support, attempting to use one of
the patterns ~g, ~G, ~k, ~V fails with the cryptic (pun intended)
error message:
  "error: unknown op %d (report this error)."

Since the op is known, just not enabled, and furthermore is an opaque
number, it doesn't help the user understand what the problem is.

Change the code to instead print out the pattern operator character
that is disabled:
  "Pattern operator '~%c' is disabled."

Almost all callers of mutt_pattern_exec() expect it to return 0 or 1.
Returning -1 in the event of a disabled or bad op will lead to
undesirable "matching" behavior.  Change to return 0 in the event of
one of these errors, since none of the callers do (or can) handle the
error in a meaningful way.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/434b55ee1c3aa2ae25731ba4a85a92f2b7ec7879
Co-authored-by: Richard Russon <rich@flatcap.org>