Fix edit-message false-modified bug
The file was being stat() before the truncate, and
mutt_file_decrease_mtime() only decreases if the sb.st_mtime is equal to
the current time.  That means operations crossing the second boundary:
  time x:   stat
  time x+1: truncate, decrease_mtime
would result in a false-modified bug if no changes were made during
editing.

The mutt_file_decrease_mtime() and callers need to be cleaned up for better
error handling, but I will do that in master.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/30340e5077d972e3b047bdcc28617f0fe5a8b97b
Co-authored-by: Richard Russon <rich@flatcap.org>