Add mitigation against DoS from thousands of parts
A demonstration attack using a million tiny parts will freeze Mutt for
several minutes.  This is actually better than some other mail
software, but can still be a problem at large levels.

For now, set it to a very conservative 5000, but this can be adjusted
up(or down) if necessary.

Declare the previous stack-limit max depth as a constant too, and
decrease it down to 50.  Change the handler to return non-fatal "1" on
reaching the limit.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/c72f740aa7c80c0a79775628e62daa2a43357cd5
Co-authored-by: Richard Russon <rich@flatcap.org>