Create $copy_decode_weed, $pipe_decode_weed, $print_decode_weed
$weed covers a bit too much functionality.  It makes sense for it to
control the display of messages in the pager, along with forwarding
and replying.  But(in my opinion) piping and copy/save-decode are
something users might like to set differently without affecting the
rest of Mutt's behavior.

$copy_decode_weed affects the <decode-copy> and <decode-save>
functions.  I've defaulted this *unset*, despite that it breaks
backward compatibility, because I believe the header weeding is
surprising to users and loses important information when copy/saving a
message.

$pipe_decode_weed affects <pipe-message> when $pipe_decode is set.
I've default this *set* for backward compatibility.

$print_decode_weed affects <print-message> when $print_decode is set.
I've default this *set* for backward compatibility.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/25f82cf6031c9c37696590924858cf553f51596f
Co-authored-by: Richard Russon <rich@flatcap.org>