improve(notmuch): don't abort sync on error
Instead of aborting the sync if there's an error on the underlying file
system, make a note of the error and display it to the user after sync
finishes.

Current behavior aborts the sync if the underlying file disappears. The
result is users unable to exit the current mailbox without either:

 1) re-opening neomutt.
 2) disabling mailbox writes and re-opening the mailbox.

Both of these methods discard the writes, which led to the conflict in
the first place. The user experience is degraded since there's no way to
recover from the write issues. Displaying the warning message provides a
better user experience that doesn't block them but makes them aware of
the issue.