Fix mutt_view_attachment() to unstuff format=flowed
This is used in receive mode for <view-text> and <view-mailcap>.
I don't think those are accessible from the compose menu, but the code
allows it.  So fix up the send-case too.

Since we may need to unstuff the send-mode file, make a copy instead
of symlinking.  Also clean up the code to just sanitize the filename
regardless.

Upstream-commit: https://gitlab.com/muttmua/mutt/commit/0600ebae1b31558f1d038da7781f6931e2c10b5d
Co-authored-by: Richard Russon <rich@flatcap.org>